# docs/openapi/api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "后台管理接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/user/token": {
      "get": {
        "summary": "生成访问令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "用户管理"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/user/kirobus/tokens": {
      "get": {
        "summary": "列出 Kiro.bus 管理的令牌",
        "deprecated": false,
        "description": "按当前认证用户、认证用户分组及 owner_system=kirobus 三重约束分页查询；响应不包含令牌 key。",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/KiroBusTokenPage"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionAuth1": [],
            "NewApiUser1": []
          },
          {
            "AccessToken1": [],
            "NewApiUser1": []
          }
        ]
      }
    },
    "/api/user/kirobus/tokens/{id}": {
      "get": {
        "summary": "获取 Kiro.bus 管理的令牌",
        "deprecated": false,
        "description": "仅返回当前认证用户、认证用户分组及 owner_system=kirobus 共同拥有的令牌；响应不包含令牌 key。",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/KiroBusToken"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "令牌不存在或不属于当前 owner scope"
          }
        },
        "security": [
          {
            "SessionAuth1": [],
            "NewApiUser1": []
          },
          {
            "AccessToken1": [],
            "NewApiUser1": []
          }
        ]
      },
      "delete": {
        "summary": "删除 Kiro.bus 管理的令牌",
        "deprecated": false,
        "description": "仅软删除当前认证用户、认证用户分组及 owner_system=kirobus 共同拥有的令牌。",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          },
          "404": {
            "description": "令牌不存在或不属于当前 owner scope"
          }
        },
        "security": [
          {
            "SessionAuth1": [],
            "NewApiUser1": []
          },
          {
            "AccessToken1": [],
            "NewApiUser1": []
          }
        ]
      }
    },
...
    "/api/token/": {
      "get": {
        "summary": "获取所有令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "post": {
        "summary": "创建令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "put": {
        "summary": "更新令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Token"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/token/{id}": {
      "get": {
        "summary": "获取指定令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      },
      "delete": {
        "summary": "删除令牌",
        "deprecated": false,
        "description": "🔐 需要登录（User权限）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/usage/token/": {
      "get": {
        "summary": "获取令牌使用情况",
        "deprecated": false,
        "description": "🔑 需要令牌认证（TokenAuth）",
        "tags": [
          "令牌管理"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
    "/api/log/token": {
      "get": {
        "summary": "通过令牌获取日志",
        "deprecated": false,
        "description": "🔓 无需鉴权（通过令牌查询）",
        "tags": [
          "日志"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "headers": {}
          }
        },
        "security": [
          {
            "Combination343": []
          },
          {
            "Combination1243": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Token": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "expired_time": {
            "type": "integer"
          },
          "remain_quota": {
            "type": "integer"
          },
          "unlimited_quota": {
            "type": "boolean"
          },
          "model_limits_enabled": {
            "type": "boolean"
          },
          "model_limits": {
            "type": "string"
          },
          "group": {
            "type": "string"
          },
          "owner_system": {
            "type": "string"
          },
          "bus_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "trip_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "seat_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "ride_order_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "client_key_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "api_route_profile_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "quota_unit": {
            "type": "string",
            "nullable": true
          },
          "rpm_limit": {
            "type": "integer",
            "nullable": true
          },
          "concurrency_limit": {
            "type": "integer",
            "nullable": true
          },
          "conversion_revision": {
            "type": "string",
            "nullable": true
          },
          "idempotency_key": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "KiroBusToken": {
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "status",
          "created_time",
          "accessed_time",
          "expired_time",
          "remain_quota",
          "used_quota",
          "unlimited_quota",
          "model_limits_enabled",
          "model_limits",
          "group",
          "owner_system"
        ],
        "properties": {
          "id": { "type": "integer" },
          "user_id": { "type": "integer" },
          "name": { "type": "string" },
          "status": { "type": "integer", "enum": [1, 2, 3, 4] },
          "created_time": { "type": "integer", "format": "int64" },
          "accessed_time": { "type": "integer", "format": "int64" },
          "expired_time": { "type": "integer", "format": "int64" },
          "remain_quota": { "type": "integer" },
          "used_quota": { "type": "integer" },
          "unlimited_quota": { "type": "boolean" },
          "model_limits_enabled": { "type": "boolean" },
          "model_limits": { "type": "string" },
          "group": { "type": "string" },
          "owner_system": { "type": "string", "enum": ["kirobus"] },
          "bus_id": { "type": "integer", "format": "int64", "nullable": true },
          "trip_id": { "type": "integer", "format": "int64", "nullable": true },
          "seat_id": { "type": "integer", "format": "int64", "nullable": true },
          "ride_order_id": { "type": "integer", "format": "int64", "nullable": true },
          "client_key_id": { "type": "integer", "format": "int64", "nullable": true },
          "api_route_profile_id": { "type": "integer", "format": "int64", "nullable": true },
          "quota_unit": { "type": "string", "nullable": true },
          "rpm_limit": { "type": "integer", "nullable": true },
          "concurrency_limit": { "type": "integer", "nullable": true },
          "conversion_revision": { "type": "string", "nullable": true },
          "idempotency_key": { "type": "string", "nullable": true }
        },
        "description": "Kiro.bus service Token metadata. The key field is intentionally absent."
      },
      "KiroBusTokenPage": {
        "type": "object",
        "required": ["page", "page_size", "total", "items"],
        "properties": {
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total": { "type": "integer" },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/KiroBusToken"
            }
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "SessionAuth1": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken1": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser1": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination222": {
        "group": [
          {
            "id": 573666
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
      "Combination1122": {
        "group": [
          {
            "id": 573667
          },
          {
            "id": 573668
          }
        ],
        "type": "combination"
      },
...
  },
  "servers": [],
  "security": [
    {
      "Combination343": []
    },
    {
      "Combination1243": []
    }
  ]
}